# docs/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Personal Style Network API",
    "version": "1.0.0",
    "description": "API cho ứng dụng Personal Style Network",
    "contact": {
      "name": "PSN Team",
      "email": "support@psn.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3052/v1/api",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/recommendations/daily": {
      "get": {
        "summary": "Lấy đề xuất hàng ngày",
        "tags": [
          "Recommendations"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Đề xuất hàng ngày"
          },
          "401": {
            "description": "Không được ủy quyền"
          }
        }
      }
    },
    "/recommendations": {
      "post": {
        "summary": "Tạo đề xuất mới",
        "tags": [
          "Recommendations"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "type"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "daily",
                      "event",
                      "weather",
                      "season",
                      "style",
                      "occasion"
                    ]
                  },
                  "context": {
                    "type": "object",
                    "properties": {
                      "date": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "weather": {
                        "type": "object",
                        "properties": {
                          "temperature": {
                            "type": "number"
                          },
                          "condition": {
                            "type": "string"
                          },
                          "humidity": {
                            "type": "number"
                          }
                        }
                      },
                      "occasion": {
                        "type": "string"
                      },
                      "location": {
                        "type": "string"
                      },
                      "promptText": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tạo đề xuất thành công"
          },
          "400": {
            "description": "Dữ liệu không hợp lệ"
          },
          "401": {
            "description": "Không được ủy quyền"
          }
        }
      },
      "get": {
        "summary": "Lấy danh sách đề xuất của người dùng",
        "tags": [
          "Recommendations"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Số trang"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            },
            "description": "Số lượng mỗi trang"
          },
          {
            "in": "query",
            "name": "type",
            "schema": {
              "type": "string",
              "default": 20,
              "enum": [
                "daily",
                "event",
                "weather",
                "season",
                "style",
                "occasion"
              ]
            },
            "description": "Lọc theo loại đề xuất"
          }
        ],
        "responses": {
          "200": {
            "description": "Danh sách đề xuất"
          },
          "401": {
            "description": "Không được ủy quyền"
          }
        }
      }
    },
    "/recommendations/{id}": {
      "get": {
        "summary": "Lấy thông tin chi tiết đề xuất",
        "tags": [
          "Recommendations"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID của đề xuất"
          }
        ],
        "responses": {
          "200": {
            "description": "Thông tin chi tiết đề xuất"
          },
          "401": {
            "description": "Không được ủy quyền"
          },
          "404": {
            "description": "Không tìm thấy đề xuất"
          }
        }
      }
    },
    "/recommendations/{id}/feedback": {
      "post": {
        "summary": "Cập nhật phản hồi người dùng cho đề xuất",
        "tags": [
          "Recommendations"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID của đề xuất"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "rating"
                ],
                "properties": {
                  "rating": {
                    "type": "number",
                    "minimum": 1,
                    "maximum": 5
                  },
                  "comment": {
                    "type": "string"
                  },
                  "selectedOutfitId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cập nhật phản hồi thành công"
          },
          "400": {
            "description": "Dữ liệu không hợp lệ"
          },
          "401": {
            "description": "Không được ủy quyền"
          },
          "404": {
            "description": "Không tìm thấy đề xuất"
          }
        }
      }
    },
...
  },
...
}